more progress

+++ Services/TodoService.cs
@@ -13,25 +13,26 @@ namespace api.Services
     public class TodoService
     {
         private readonly DataContext _context;
+
         public TodoService(DataContext context)
         {
             _context = context;
         }
 
-        public async Task<List<TodosModel>> GetTodosAsync () => await _context.TodosInfo.ToListAsync();
-        private async Task<TodosModel> GetTodoByIdAsync(int id)
+        public async Task<List<TodosModel>> GetTodos () => await _context.TodosInfo.ToListAsync();
+        public async Task<TodosModel> GetTodoById(int id)
         {
             return await _context.TodosInfo.FindAsync(id);
         }
-        public async Task<bool> AddTodoAsync(TodosModel todo)
+        public async Task<bool> AddTodo(TodosModel todo)
         { 
              await _context.TodosInfo.AddAsync(todo);
             return await _context.SaveChangesAsync() != 0;
         }
 
-        public async Task<bool> EditTodoAsync(TodosModel todo)
+        public async Task<bool> EditTodo(TodosModel todo)
         {
-            var todoToEdit = await GetTodoByIdAsync(todo.Id);
+            var todoToEdit = await GetTodoById(todo.Id);
 
             if(todoToEdit == null) return false;
 
@@ -39,15 +40,20 @@ namespace api.Services
             todoToEdit.Difficulty = todo.Difficulty;
             todoToEdit.Completed = todo.Completed;
             todoToEdit.Deleted = todo.Deleted;
-            
 
             _context.TodosInfo.Update(todoToEdit);
             return await _context.SaveChangesAsync() != 0;
         }
 
-public async Task<bool> HardDeleteTodo(TodosModel todo)
+public async Task<bool> HardDeleteTodo(int id)
 {
+    var todo = await _context.TodosInfo.FindAsync(id);
+
+    if (todo == null)
+        return false;
+
     _context.TodosInfo.Remove(todo);
+
     return await _context.SaveChangesAsync() > 0;
 }
 
@@ -67,7 +73,7 @@ public async Task<bool> HardDeleteTodo(TodosModel todo)
 
         public async Task<List<TodosModel>> GetTodoByUserIdAsync(int id) => await _context.TodosInfo.Where(todo => todo.UserId == id).ToListAsync();
 
-        public async Task<List<TodosModel>> GetIncompleteTodosAsync() => await _context.TodosInfo.Where(todo => todo.Completed == false).ToListAsync();
+        public async Task<List<TodosModel>> GetIncompleteTodos() => await _context.TodosInfo.Where(todo => todo.Completed == false).ToListAsync();
     // }
 
     }
